requirements.txt
# ── Web framework ─────────────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1

# ── Database ──────────────────────────────────────────────────────
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# ── Validation & settings ─────────────────────────────────────────
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.6.1

# ── Authentication ────────────────────────────────────────────────
bcrypt==4.2.1
PyJWT==2.10.1

# ── HTTP client (for Ollama API calls) ────────────────────────────
httpx==0.28.1

# ── Form data support ─────────────────────────────────────────────
python-multipart==0.0.18

# ── SMS / Notifications ───────────────────────────────────────────
twilio==9.4.0
phonenumbers==8.13.50

# ── Scheduler (daily reminders) ───────────────────────────────────
apscheduler==3.10.4
pytz==2024.2

# ── Testing ───────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
